pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wma-cross-alerts"
version = "0.1.0"
description = "Sistema de alertas basado en Golden Cross WMA"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "matplotlib>=3.8",
    "yfinance>=0.2",
    "pyyaml>=6.0",
    "requests>=2.31",
    "python-dateutil>=2.8",
    "pytz>=2023.3",
    "python-dotenv>=1.0",
    "lxml>=6.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
